- DEC Alpha port: make things 64-bit clean (again)

+++ platform/emulator/libdp/table.hh
@@ -41,7 +41,7 @@
 #include "genhashtbl.hh"
 #include "perdio.hh"
 
-typedef long Credit;
+typedef int Credit;
 
 #ifdef DEBUG_CHECK
 Bool withinBorrowTable(int i);
